Flytter litt rundt på kode
