+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/eøs/differanseberegning/domene/Intervall.kt
@@ -1,0 +1,8 @@
+package no.nav.familie.ba.sak.kjerne.eøs.differanseberegning.domene
+
+enum class Intervall {
+    ÅRLIG,
+    KVARTALSVIS,
+    MÅNEDLIG,
+    UKENTLIG;
+}
